chore: remove deprecated actions-rs/toolchain github actions and replace with dtolnay/rust-toolchain
All of the actions-rs actions are effectively un-maintained now. for the toolchain action, dtolnay/rust-toolchain
appears to be the community's choice for drop-in replacement.

Additionally, in places where we install a nightly toolchain, install the stable cargo as well. There were several
actions where I observed errors because the cargo installed in the image did not support rust2024 yet